# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Plataforma Practicas API
  version: 1.0.0
  description: |
    API REST para gestión de blogs, comentarios, multimedia y autenticación.
    
    **Interfaz Swagger UI disponible en:** `/api-docs`
servers:
  - url: http://localhost:3000/api
paths:
...
  /comentario:
    get:
      summary: Obtener todos los comentarios
      responses:
        '200':
          description: Lista de comentarios
          content:
            application/json:
              example:
                - id: 1
                  contenido: "Comentario de ejemplo"
                  blogPostId: 1
                  autorId: 2
                  autorTipo: usuario
                  parentId: null
                  createdAt: "2024-07-08T12:00:00Z"
                  updatedAt: "2024-07-08T12:00:00Z"
    post:
      summary: Crear comentario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contenido:
                  type: string
                  example: "Nuevo comentario"
                blogPostId:
                  type: integer
                  example: 1
                autorId:
                  type: integer
                  example: 2
                autorTipo:
                  type: string
                  example: usuario
                parentId:
                  type: integer
                  example: null
      responses:
        '201':
          description: Comentario creado
          content:
            application/json:
              example:
                mensaje: Comentario creado
                comentario:
                  id: 2
                  contenido: "Nuevo comentario"
                  blogPostId: 1
                  autorId: 2
                  autorTipo: usuario
                  parentId: null
                  createdAt: "2024-07-08T12:00:00Z"
                  updatedAt: "2024-07-08T12:00:00Z"
        '400':
          description: Error de validación
          content:
            application/json:
              example:
                error: Faltan campos requeridos
  /comentario/{id}:
    get:
      summary: Obtener comentario por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comentario encontrado
          content:
            application/json:
              example:
                id: 1
                contenido: "Comentario de ejemplo"
                blogPostId: 1
                autorId: 2
                autorTipo: usuario
                parentId: null
                createdAt: "2024-07-08T12:00:00Z"
                updatedAt: "2024-07-08T12:00:00Z"
        '404':
          description: No encontrado
          content:
            application/json:
              example:
                error: Comentario no encontrado
    put:
      summary: Actualizar comentario
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contenido:
                  type: string
                  example: "Comentario actualizado"
      responses:
        '200':
          description: Comentario actualizado
          content:
            application/json:
              example:
                mensaje: Comentario actualizado
                comentario:
                  id: 1
                  contenido: "Comentario actualizado"
                  blogPostId: 1
                  autorId: 2
                  autorTipo: usuario
                  parentId: null
                  createdAt: "2024-07-08T12:00:00Z"
                  updatedAt: "2024-07-08T12:10:00Z"
        '403':
          description: No autorizado
          content:
            application/json:
              example:
                error: No autorizado para editar este comentario
        '404':
          description: No encontrado
          content:
            application/json:
              example:
                error: Comentario no encontrado
    delete:
      summary: Eliminar comentario
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comentario eliminado
          content:
            application/json:
              example:
                mensaje: Comentario eliminado correctamente
        '403':
          description: No autorizado
          content:
            application/json:
              example:
                error: No autorizado para eliminar este comentario
        '404':
          description: No encontrado
          content:
            application/json:
              example:
                error: Comentario no encontrado
...
  /empresa:
    get:
      summary: Obtener todas las empresas
      responses:
        '200':
          description: Lista de empresas
          content:
            application/json:
              example:
                - id: 1
                  nombre: Empresa A
                  descripcion: Descripción de la empresa A
                  createdAt: "2024-07-08T12:00:00Z"
                  updatedAt: "2024-07-08T12:00:00Z"
    post:
      summary: Crear empresa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: Empresa A
                descripcion:
                  type: string
                  example: Descripción de la empresa A
      responses:
        '201':
          description: Empresa creada
          content:
            application/json:
              example:
                mensaje: Empresa creada
                empresa:
                  id: 2
                  nombre: Empresa A
                  descripcion: Descripción de la empresa A
                  createdAt: "2024-07-08T12:00:00Z"
                  updatedAt: "2024-07-08T12:00:00Z"
  /empresa/{id}:
    get:
      summary: Obtener empresa por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Empresa encontrada
          content:
            application/json:
              example:
                id: 1
                nombre: Empresa A
                descripcion: Descripción de la empresa A
                createdAt: "2024-07-08T12:00:00Z"
                updatedAt: "2024-07-08T12:00:00Z"
        '404':
          description: No encontrado
          content:
            application/json:
              example:
                error: Empresa no encontrada
    put:
      summary: Actualizar empresa
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: Empresa A
                descripcion:
                  type: string
                  example: Descripción de la empresa A
      responses:
        '200':
          description: Empresa actualizada
          content:
            application/json:
              example:
                mensaje: Empresa actualizada
                empresa:
                  id: 1
                  nombre: Empresa A
                  descripcion: Descripción de la empresa A
                  createdAt: "2024-07-08T12:00:00Z"
                  updatedAt: "2024-07-08T12:10:00Z"
        '404':
          description: No encontrado
          content:
            application/json:
              example:
                error: Empresa no encontrada
    delete:
      summary: Eliminar empresa
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Empresa eliminada
          content:
            application/json:
              example:
                mensaje: Empresa eliminada correctamente
        '404':
          description: No encontrado
          content:
            application/json:
              example:
                error: Empresa no encontrada
...